posix: fix signedness issue with ioctls
The Linux definition of ioctl takes an unsigned long for the 'command'/'request' parameter - and Pedigree has been using an int. For some ioctls (e.g. those above 0x80000000) this breaks as the cast from int to size_t sign-extends and as a result the ioctl call doesn't match any known commands.

Using an unsigned type end-to-end fixes this up.

+++ src/modules/subsys/posix/DevFs.cc
@@ -195,7 +195,7 @@ uint64_t RtcFile::writeBytewise(
     return 0;
 }
 
-bool RtcFile::supports(const int command) const
+bool RtcFile::supports(const size_t command) const
 {
     // read/set time
     return true;
@@ -203,7 +203,7 @@ bool RtcFile::supports(const int command) const
     // static_cast<size_t>(command) == 0x4024700aUL;
 }
 
-int RtcFile::command(const int command, void *buffer)
+int RtcFile::command(const size_t command, void *buffer)
 {
     NOTICE("RtcFile: command " << Hex << command << " with buffer " << buffer);
     return 0;
@@ -278,12 +278,12 @@ uintptr_t FramebufferFile::readBlock(uint64_t location)
            location;
 }
 
-bool FramebufferFile::supports(const int command) const
+bool FramebufferFile::supports(const size_t command) const
 {
     return (PEDIGREE_FB_CMD_MIN <= command) && (command <= PEDIGREE_FB_CMD_MAX);
 }
 
-int FramebufferFile::command(const int command, void *buffer)
+int FramebufferFile::command(const size_t command, void *buffer)
 {
     if (!m_pGraphicsParameters)
     {
